+++ TablePro/Theme/RegistryThemeMeta.swift
@@ -1,0 +1,16 @@
+import Foundation
+
+struct RegistryThemeMeta: Codable {
+    var installed: [InstalledRegistryTheme]
+
+    init(installed: [InstalledRegistryTheme] = []) {
+        self.installed = installed
+    }
+}
+
+struct InstalledRegistryTheme: Codable, Identifiable {
+    let id: String
+    let registryPluginId: String
+    let version: String
+    let installedDate: Date
+}